% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{BuildConsensus-package}
\alias{BuildConsensus}
\alias{BuildConsensus-package}
\title{Build a Consensus Classification from Multiple Classification Systems}
\description{
This package implements methods for building a consensus molecular 
classification using existing classification systems.
}
\note{
This is a contribution from the Tumor Identity Cards (CIT) program founded by the 
'Ligue Nationale Contre le Cancer' (France): \url{http://cit.ligue-cancer.net}. 
For any question please contact \url{CITR@ligue-cancer.net}
}
\seealso{
\link[BuildConsensus]{consensus.MCL}, \link[BuildConsensus]{consensus.CIT},
\link[BuildConsensus]{consensus.COCA}
}
\author{
Aurelien de Reynies, Aurelie Kamoun
}
