% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineConsensus.R
\name{consensus.COCA}
\alias{consensus.COCA}
\title{Build a consensus classification using the COCA method}
\usage{
consensus.COCA(annot.class, outdir, maxK = 8, reps = 100, seed = 12,
  innerLinkage = "ward.D2", finalLinkage = "ward.D2",
  distance = "euclidean", ...)
}
\arguments{
\item{annot.class}{a dataframe of samples annotated according to the several 
classification systems to compare.}

\item{outdir}{the path to the directory where to store plots and results. 
A new directory will be created if the supplied path does not exist.}

\item{maxK}{an integer value that aximum cluster number to be evaluated.}

\item{reps}{an integer value that represents the number of repetitions to be performed.}

\item{seed}{an integer. Used for result reproducibility. See \link{set.seed}}

\item{innerLinkage}{a linkage method for constructing the hierarchical clustering. 
See \link[stats]{hclust} for linkage methods.}

\item{finalLinkage}{a linkage method for constructing the final clustering. 
See \link[stats]{hclust} for linkage methods.}

\item{distance}{a distance metric to evaluate the samples. See \link[stats]{dist} for methods.}

\item{...}{other parameters passed to \link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
\description{
Compare and cluster various classification systems using a "Cluster of Cluster" approach
}
\examples{
\dontrun{
 #-- Using bladder cancer classes as example
 library(BuildConsensus)
 data(blca_class)
 
 coca_res <- consensus.coca(blca_class[sample(nrow(blca_class), 500),], outdir = "coca_res", maxK = 5)
}

}
\author{
Aurelie Kamoun
}
